requirements.txt
streamlit
openai
faiss-cpu
PyPDF2
tiktoken